chore: refactor syncing of params and stores
